requirements.txt
numpy==1.19.2
python-dateutil==2.8.1
python-dotenv==0.14.0
discord.py==1.4.1
pandas==1.1.2